backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
reportlab==4.0.9
pandas==2.2.0
qrcode==7.4.2
pillow==12.1.1
pydantic[email]==2.5.3
pydantic-settings==2.1.0
PyPDF2==3.0.1
aiosmtplib==3.0.1
google-genai==0.3.0
openai==1.14.0
dnspython>=2.6.0
slowapi==0.1.9
